+++ core/lib/dal/migrations/20260713000000_l1_batch_cycle_stats.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE l1_batch_cycle_stats;